+++ Models/UserTopic.cs
@@ -5,7 +5,7 @@ namespace WebHelloWorld.Models
     public class UserTopic
     {
         [Column("user_id")]
-        public int UserId { get; set; }
+        public Int64 UserId { get; set; }
         public User User { get; set; }
         
         [Column("topic_id")]
